+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ---------
 - Fix typo in russian translation
+- Fix width errors where drop-in was not aligned with other elements on merchant page
 
 1.1.0
 ----------