+++ CHANGELOG.md
@@ -7,6 +7,8 @@ UNRELEASED
 - Fix typo in russian translation
 - Update browser detection library to 1.4.0
 - Fix width errors where Drop-in was not aligned with other elements on merchant page
+- Add ability to style PayPal button
+- Fail early if PayPal creation errors
 - Add hungarian translation
 
 1.1.0