+++ CHANGELOG.md
@@ -9,6 +9,8 @@ UNRELEASED
 - Fix width errors where Drop-in was not aligned with other elements on merchant page
 - Add ability to style PayPal button
 - Fail early if PayPal creation errors
+- Upgrade braintree-web to v3.17.0
+- Upgrade checkout.js to 4.0.78
 
 1.1.0
 ----------